remove test component, abstract socketplugin from store file, cleanup/comment store, cleanup App
