+++ packages/vuetron-app/src/App.vue
@@ -11,8 +11,6 @@
   </b-container>
 </template>
 
-
-
 <script>
   import Navbar from './components/Navbar.vue';
   export default {
